refactor: 테스트 기반 command 코드 수정

+++ src/main/kotlin/kr/co/shoppingcart/cart/domain/template/TemplateUseCase.kt
@@ -30,7 +30,7 @@ class TemplateUseCase(
 
     fun getByIdAndUserId(getTemplateByIdAndUserIdCommand: GetTemplateByIdAndUserIdCommand) =
         templateRepository.getByIdAndUserId(
-            getTemplateByIdAndUserIdCommand.id.toLong(),
+            getTemplateByIdAndUserIdCommand.id,
             getTemplateByIdAndUserIdCommand.userId,
         ) ?: throw CustomException.responseBody(ExceptionCode.E_403_000)
